+++ teleprox/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.2.2"
+__version__ = "2.3.0"
 
 from .client import RPCClient, RemoteCallException, Future
 from .server import RPCServer